+++ web/projects/openinsurance/system/manageInsurars.jsp
@@ -25,10 +25,10 @@
     private String addCategory(int iTotal, String sCatName, String sCatLabel, String sCatPatientPercentage, String sWebLanguage){
         return "<tr id='rowCategory"+iTotal+"' class='"+(iTotal%2==0?"list":"list1")+"'>"+
                 "<td>"+
-                 "<a href='javascript:deleteCategory(rowCategory"+iTotal+");'>"+
+                 "<a href='#' onclick='deleteCategory(rowCategory"+iTotal+");'>"+
                   "<img src='"+sCONTEXTPATH+"/_img/icon_delete.gif' alt='"+getTranNoLink("Web","delete",sWebLanguage)+"' border='0'>"+
                  "</a>&nbsp;"+
-                 "<a href='javascript:editCategory(rowCategory"+iTotal+");'>"+
+                 "<a href='#' onclick='editCategory(rowCategory"+iTotal+");'>"+
                   "<img src='"+sCONTEXTPATH+"/_img/icon_edit.gif' alt='"+getTranNoLink("Web","edit",sWebLanguage)+"' border='0'>"+
                  "</a>&nbsp;"+
                 "</td>"+
@@ -179,7 +179,7 @@
         sBackFunction = "doBackToMenu();";
     }
 %>
-<form id="transactionForm" name="transactionForm" method="post" <%=sOnKeyDown%>>
+<form id="transactionForm" name="transactionForm" method="post" <%=sOnKeyDown%> onclick="setSaveButton(event);" onkeyup="setSaveButton(event);">
     <%-- hidden fields --%>
     <input type="hidden" name="Action">
     <input type="hidden" name="FindInsurarId">
@@ -248,7 +248,7 @@
 	                                    %>
 	                                        <tr class="list<%=sClass%>" onmouseover="this.style.cursor='hand';" onmouseout="this.style.cursor='default';">
 	                                            <td>
-	                                                <a href="javascript:deleteInsurar('<%=insurar.getUid()%>');"><img src='<c:url value="/_img/icon_delete.gif"/>' border='0' alt="<%=sTranDelete%>"></a>
+	                                                <a href="#" onclick="deleteInsurar('<%=insurar.getUid()%>');"><img src='<c:url value="/_img/icon_delete.gif"/>' border='0' alt="<%=sTranDelete%>"></a>
 	                                            </td>
 	                                            <td class="hand" onClick="editInsurar('<%=insurar.getUid()%>');"><%=checkString(insurar.getName())%></td>
 	                                            <td class="hand" onClick="editInsurar('<%=insurar.getUid()%>');"><%=checkString(insurar.getOfficialName())%></td>
@@ -398,7 +398,7 @@
                         	}
                         %>
                         <td class="admin2">
-                            <input type="checkbox" name="EditInsurarExtra" <%=sExtraInsurar %>/>
+                            <input type="checkbox" class="text" name="EditInsurarExtra" <%=sExtraInsurar %>/>
                         </td>
                     </tr>
                     <%-- SELECTED CATEGORIES --%>
@@ -545,7 +545,7 @@
 
   <%-- DELETE INSURAR --%>
   function deleteInsurar(sInsurarUid){
-      if(yesnoDeleteDialog()){
+    if(yesnoDialog("Web","areYouSureToDelete")){
       transactionForm.EditInsurarId.value = sInsurarUid;
       transactionForm.Action.value = "delete";
       transactionForm.submit();
@@ -666,8 +666,8 @@
         row.insertCell(i);
       }
 
-      row.cells[0].innerHTML = "<a href='javascript:deleteCategory(rowCategory"+iIndexCategories+");'><img src='<%=sCONTEXTPATH%>/_img/icon_delete.gif' alt='<%=getTranNoLink("Web","delete",sWebLanguage)%>' border='0'></a>&nbsp;"+
-                               "<a href='javascript:editCategory(rowCategory"+iIndexCategories+");'><img src='<%=sCONTEXTPATH%>/_img/icon_edit.gif' alt='<%=getTranNoLink("Web","edit",sWebLanguage)%>' border='0'></a>&nbsp;";
+      row.cells[0].innerHTML = "<a href='#' onclick='deleteCategory(rowCategory"+iIndexCategories+");'><img src='<%=sCONTEXTPATH%>/_img/icon_delete.gif' alt='<%=getTranNoLink("Web","delete",sWebLanguage)%>' border='0'></a>&nbsp;"+
+                               "<a href='#' onclick='editCategory(rowCategory"+iIndexCategories+");'><img src='<%=sCONTEXTPATH%>/_img/icon_edit.gif' alt='<%=getTranNoLink("Web","edit",sWebLanguage)%>' border='0'></a>&nbsp;";
       row.cells[1].innerHTML = catName;
 
       // remove quotes from label
@@ -729,8 +729,8 @@
 
       <%-- update table object --%>
       var row = tblCategories.rows[editCategoryRowid.rowIndex];
-      row.cells[0].innerHTML = "<a href='javascript:deleteCategory("+editCategoryRowid.id+")'><img src='<%=sCONTEXTPATH%>/_img/icon_delete.gif' alt='<%=getTran("web","delete",sWebLanguage)%>' border='0'></a> "+
-                               "<a href='javascript:editCategory("+editCategoryRowid.id+")'><img src='<%=sCONTEXTPATH%>/_img/icon_edit.gif' alt='<%=getTran("web","edit",sWebLanguage)%>' border='0'></a>";
+      row.cells[0].innerHTML = "<a href='#' onclick='deleteCategory("+editCategoryRowid.id+")'><img src='<%=sCONTEXTPATH%>/_img/icon_delete.gif' alt='<%=getTran("web","delete",sWebLanguage)%>' border='0'></a> "+
+                               "<a href='#' onclick='editCategory("+editCategoryRowid.id+")'><img src='<%=sCONTEXTPATH%>/_img/icon_edit.gif' alt='<%=getTran("web","edit",sWebLanguage)%>' border='0'></a>";
 
       row.cells[1].innerHTML = transactionForm.EditCategoryName.value;
       row.cells[2].innerHTML = transactionForm.EditCategoryLabel.value;
@@ -781,7 +781,7 @@
       else{
         if(patientCount==0){
           <%-- no patients in this category, so delete it --%>
-          if(yesnoDeleteDialog()){
+          if(yesnoDialog("Web","areYouSureToDelete")){
             sCategories = deleteRowFromArrayString(sCategories,rowid.id);
             initCategoriesArray(sCategories);
             tblCategories.deleteRow(rowid.rowIndex);
@@ -794,8 +794,11 @@
           <%-- some patients in this category, so ask again before deleting it --%>
           var msg = "<%=getTranNoLink("web.manage","patientsInThisCategory",sWebLanguage)%><br><%=getTranNoLink("web","areyousuretodelete",sWebLanguage)%>";
           msg = replaceAll(msg,"#patientCount#",patientCount);
-        
-          if(yesnoDialogDirectText(msg)){
+          var popupUrl = "<%=sCONTEXTPATH%>/_common/search/yesnoPopup.jsp?ts=<%=getTs()%>&labelValue="+msg;
+          var modalities = "dialogWidth:266px;dialogHeight:143px;center:yes;scrollbars:no;resizable:no;status:no;location:no;";
+          var answer = (window.showModalDialog)?window.showModalDialog(popupUrl,'',modalities):window.confirm('<%=getTranNoLink("web","areyousure",sWebLanguage)%>');
+
+          if(answer==1){
             categoriesToDeleteOnSave+= categoryName+"$";
 
             sCategories = deleteRowFromArrayString(sCategories,rowid.id);
@@ -810,7 +813,7 @@
     }
     else{
       <%-- category that has not been saved yet, so delete it --%>
-      if(yesnoDeleteDialog()){
+      if(yesnoDialog("Web","areYouSureToDelete")){
         sCategories = deleteRowFromArrayString(sCategories,rowid.id);
         initCategoriesArray(sCategories);
         tblCategories.deleteRow(rowid.rowIndex);
@@ -898,7 +901,7 @@
   <%-- DELETE ALL CATEGORIES --%>
   function deleteAllCategories(){
     if(tblCategories.rows.length > 1){
-        if(yesnoDeleteDialog()){
+      if(yesnoDialog("Web","areYouSureToDelete")){
         deleteAllCategoriesNoConfirm();
       }
     }
